add note concept application
