+++ public/index.html
@@ -1,41 +1,49 @@
 <!doctype html>
 <html lang="en">
-  <head>
-    <title>CS4241 Assignment 2</title>
+<head>
+    <title>René Borner CS4241 Assignment 2</title>
+    <link rel="stylesheet" href="css/style.css">
+    <link rel="icon" href="https://reneborner.de/favicon.ico">
     <meta charset="utf-8">
-  </head>
-  <body>
-    <form action="">
-      <input type='text' id='yourname' value="your name here">
-      <button>submit</button>
-    </form>
-  </body>
-  <script>
-
-  const submit = function( e ) {
-    // prevent default form action from being carried out
-    e.preventDefault()
-
-    const input = document.querySelector( '#yourname' ),
-          json = { yourname: input.value },
-          body = JSON.stringify( json )
-
-    fetch( '/submit', {
-      method:'POST',
-      body 
-    })
-    .then( function( response ) {
-      // do something with the reponse 
-      console.log( response )
-    })
-
-    return false
-  }
-
-  window.onload = function() {
-    const button = document.querySelector( 'button' )
-    button.onclick = submit
-  }
-
-  </script>
+    <script src="js/Chart.min.js"></script>
+</head>
+<body>
+<h1>
+    <img class="profile"
+         src="images/noteLogo.png"
+         alt="
+     avatar">
+    <br>
+    Hi I'm your notes app!
+</h1>
+<!-- maybe do something with graphJS? -->
+<form action="">
+    <input type="text" id="inputText" placeholder="Text" value="test">
+    <input type="date" id="inputDate" placeholder="Due">
+    <button>Save</button>
+</form>
+<script>
+    
+    const submit = function (e) {
+        // prevent default form action from being carried out
+        e.preventDefault()
+        const inputText = document.getElementById( 'inputText' ).value
+        const inputDate = document.getElementById( 'inputDate' ).value
+        const json = {Note: inputText, Date: inputDate}
+        const body = JSON.stringify(json)
+        fetch('/submit', {
+            method: 'POST',
+            body
+        })
+            .then(function (response) {
+                console.log(response.text())
+            })
+        return false
+    }
+    window.onload = function () {
+        const button = document.querySelector('button')
+        button.onclick = submit
+    }
+</script>
+</body>
 </html>
