--- public/js/scripts.js
@@ -1,3 +1,0 @@
-// Add some Javascript code here, to run on the front end.
-
-console.log("Welcome to assignment 2!")
